# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "合同管理系统 API",
    "version": "1.0.0",
    "description": "\n# 合同管理系统 API 文档\n\n## 概述\n\n基于契约优先（Contract-First）架构的 RESTful API，用于管理华道业务的合同、用户和部门。\n\n## 认证方式\n\n系统采用 JWT 双 Token 认证机制：\n\n- **Access Token**: 用于 API 认证，有效期 15 分钟\n- **Refresh Token**: 用于刷新 Access Token，有效期 7 天\n\n### 获取 Token\n\n调用 `POST /api/auth/login` 获取 Access Token 和 Refresh Token。\n\n### 使用 Token\n\n在请求头中添加：\n```\nAuthorization: Bearer <access_token>\n```\n\n### 刷新 Token\n\n当 Access Token 过期时，调用 `POST /api/auth/refresh` 使用 Refresh Token 获取新的 Access Token。\n\n## 权限控制\n\n### 功能权限（菜单权限）\n\n用户只能看到和访问被授权的功能模块。\n\n### 数据权限\n\n- **普通用户 (EMPLOYEE)**: 只能查看自己创建的数据\n- **管理员 (ADMIN)**: 可以查看所有数据\n\n## 错误响应\n\n所有错误响应遵循 RFC 7807 Problem Details 规范：\n\n```json\n{\n  \"type\": \"https://example.com/errors/validation-error\",\n  \"title\": \"Validation Error\",\n  \"status\": 400,\n  \"detail\": \"One or more fields failed validation\",\n  \"instance\": \"/api/users\",\n  \"errors\": [\n    {\n      \"field\": \"email\",\n      \"message\": \"Invalid email format\"\n    }\n  ]\n}\n```\n\n## 分页\n\n列表接口支持分页，通过查询参数控制：\n\n- `page`: 页码（从 1 开始）\n- `limit`: 每页数量\n\n分页响应包含元数据：\n\n```json\n{\n  \"success\": true,\n  \"data\": [...],\n  \"meta\": {\n    \"page\": 1,\n    \"limit\": 20,\n    \"total\": 100,\n    \"totalPages\": 5\n  },\n  \"timestamp\": \"2026-06-08T12:00:00Z\"\n}\n```\n    ",
    "contact": {
      "name": "API Support",
      "email": "support@huada.com"
    },
...
    }
...
  "servers": [
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/departments": {
      "get": {
        "summary": "获取部门列表",
        "description": "获取部门列表，支持树形结构",
        "tags": [
          "部门管理"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "tree",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "description": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "parentId": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "managerId": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "ACTIVE",
                              "INACTIVE"
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "code",
                          "description",
                          "parentId",
                          "managerId",
                          "status",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "post": {
        "summary": "创建部门",
        "description": "创建新部门",
        "tags": [
          "部门管理"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  },
                  "description": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "parentId": {
                    "anyOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "managerId": {
                    "anyOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "INACTIVE"
                    ]
                  },
                  "createdAt": {
                    "type": "string"
                  },
                  "updatedAt": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "code",
                  "description",
                  "parentId",
                  "managerId",
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "name": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "description": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "parentId": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "managerId": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "ACTIVE",
                            "INACTIVE"
                          ]
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "code",
                        "description",
                        "parentId",
                        "managerId",
                        "status",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限创建部门"
          },
          "409": {
            "description": "部门编码已存在"
          }
        }
      }
    },
    "/api/departments/{id}": {
      "get": {
        "summary": "获取部门详情",
        "description": "根据 ID 获取部门详情",
        "tags": [
          "部门管理"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "name": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "description": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "parentId": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "managerId": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "ACTIVE",
                            "INACTIVE"
                          ]
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "code",
                        "description",
                        "parentId",
                        "managerId",
                        "status",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "部门不存在"
          }
        }
      },
      "put": {
        "summary": "更新部门",
        "description": "更新部门信息",
        "tags": [
          "部门管理"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  },
                  "description": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "parentId": {
                    "anyOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "managerId": {
                    "anyOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "INACTIVE"
                    ]
                  },
                  "createdAt": {
                    "type": "string"
                  },
                  "updatedAt": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "code",
                  "createdAt",
                  "updatedAt"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "name": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "description": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "parentId": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "managerId": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "ACTIVE",
                            "INACTIVE"
                          ]
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "code",
                        "description",
                        "parentId",
                        "managerId",
                        "status",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限修改部门"
          },
          "404": {
            "description": "部门不存在"
          }
        }
      },
      "delete": {
        "summary": "删除部门",
        "description": "删除部门",
        "tags": [
          "部门管理"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限删除部门"
          },
          "404": {
            "description": "部门不存在"
          },
          "409": {
            "description": "部门下存在用户或子部门，无法删除"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 访问令牌认证"
      }
    }
  }
}